+++ scripts/macros/2024/classFeatures/warlock/invocations/giftOfTheProtectors.js
@@ -1,0 +1,5 @@
+import {giftOfTheProtectors as giftOfTheProtectorsLegacy} from '../../../../../legacyMacros.js';
+export let giftOfTheProtectors = {
+    ...giftOfTheProtectorsLegacy,
+    rules: 'modern'
+};